refactor: reorganize project to monorepo
