+++ packages/select/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "@pretear/select",
+  "version": "1.0.0",
+  "main": "index.js",
+  "license": "MIT",
+  "peerDependencies": {
+    "@chakra-ui/system": "2.0.0",
+    "react": ">=18.0.0"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.17.10",
+    "@chakra-ui/system": "^2.0.0",
+    "@types/react": "^18.0.9",
+    "@vitejs/plugin-react": "^1.3.2",
+    "react": "18.0.0",
+    "typescript": "^4.6.4",
+    "unplugin-swc": "^1.3.2",
+    "vite": "^2.9.9",
+    "vite-plugin-dts": "^1.1.1",
+    "vite-tsconfig-paths": "^3.4.1"
+  },
+  "dependencies": {
+    "@chakra-ui/form-control": "2.0.0",
+    "@chakra-ui/utils": "2.0.0"
+  }
+}
